Move MemoryMixin to own file
