+++ angr/storage/__init__.py
@@ -4,13 +4,12 @@ DUMMY_SYMBOLIC_READ_VALUE = 0xC0DEB4BE
 
 from .file import SimFile
 from .memory_object import SimMemoryObject
-from .memory_mixins import MemoryMixin, DefaultMemory
+from .memory_mixins import DefaultMemory
 
 
 __all__ = (
     "DUMMY_SYMBOLIC_READ_VALUE",
     "SimFile",
     "SimMemoryObject",
-    "MemoryMixin",
     "DefaultMemory",
 )
