+++ angr/storage/memory_mixins/actions_mixin.py
@@ -3,7 +3,7 @@ import claripy
 
 from angr.state_plugins.sim_action import SimActionData, SimActionObject
 from angr import sim_options as o
-from . import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class ActionsMixinHigh(MemoryMixin):
