+++ angr/storage/memory_mixins/address_concretization_mixin.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 import claripy
 
-from . import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 from angr import sim_options as options
 from angr import concretization_strategies
 from angr.sim_state_options import SimStateOptions
