+++ angr/storage/memory_mixins/conditional_store_mixin.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 import claripy
 
-from . import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class ConditionalMixin(MemoryMixin):
