+++ angr/storage/memory_mixins/convenient_mappings_mixin.py
@@ -7,7 +7,7 @@ import claripy
 from angr import sim_options as options
 from angr.utils.cowdict import ChainMapCOW
 from angr.errors import SimMemoryError, SimMemoryMissingError
-from . import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 l = logging.getLogger(name=__name__)
