+++ angr/storage/memory_mixins/keyvalue_memory_mixin.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from . import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class TypedVariable:
