+++ angr/storage/memory_mixins/name_resolution_mixin.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
+
 import claripy
 from archinfo.arch_arm import is_arm_arch
-from . import MemoryMixin
+
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 stn_map = {"st%d" % n: n for n in range(8)}
 tag_map = {"tag%d" % n: n for n in range(8)}
