+++ angr/storage/memory_mixins/paged_memory/paged_memory_mixin.py
@@ -7,7 +7,7 @@ from collections import defaultdict
 
 import claripy
 
-from angr.storage.memory_mixins import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 from angr.storage.memory_mixins.paged_memory.pages import PageType, ListPage, UltraPage, MVListPage
 from angr.errors import SimMemoryError
