+++ angr/storage/memory_mixins/paged_memory/paged_memory_multivalue_mixin.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
-from angr.storage.memory_mixins import MemoryMixin
+
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class PagedMemoryMultiValueMixin(MemoryMixin):
