+++ angr/storage/memory_mixins/paged_memory/pages/ispo_mixin.py
@@ -1,6 +1,7 @@
 # pylint:disable=arguments-differ
 from __future__ import annotations
-from angr.storage.memory_mixins import MemoryMixin
+
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class ISPOMixin(MemoryMixin):
