+++ angr/storage/memory_mixins/paged_memory/pages/permissions_mixin.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 import claripy
 
-from angr.storage.memory_mixins import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class PermissionsMixin(MemoryMixin):
