+++ angr/storage/memory_mixins/paged_memory/pages/refcount_mixin.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
-from angr.storage.memory_mixins import MemoryMixin
+
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 from angr.misc import PicklableLock
