+++ angr/storage/memory_mixins/regioned_memory/abstract_merger_mixin.py
@@ -4,8 +4,7 @@ import logging
 from collections.abc import Iterable
 from typing import Any
 
-
-from angr.storage.memory_mixins import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 l = logging.getLogger(name=__name__)
