+++ angr/storage/memory_mixins/regioned_memory/region_meta_mixin.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 import copy
 from typing import Any
 
-from angr.storage.memory_mixins import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class Segment:
