+++ angr/storage/memory_mixins/regioned_memory/regioned_address_concretization_mixin.py
@@ -6,7 +6,7 @@ import claripy
 from angr.sim_options import HYBRID_SOLVER, APPROXIMATE_FIRST
 from angr import concretization_strategies
 from angr.errors import SimMergeError, SimMemoryAddressError
-from angr.storage.memory_mixins import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 from .abstract_address_descriptor import AbstractAddressDescriptor
 from .region_data import AddressWrapper
