+++ angr/storage/memory_mixins/simplification_mixin.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
-from . import MemoryMixin
+
 from angr import sim_options as options
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class SimplificationMixin(MemoryMixin):
