+++ angr/storage/memory_mixins/slotted_memory.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 import claripy
 
-from . import MemoryMixin
-from .paged_memory.pages.ispo_mixin import ISPOMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
+from angr.storage.memory_mixins.paged_memory.pages.ispo_mixin import ISPOMixin
 from angr.errors import SimMergeError
