+++ angr/storage/memory_mixins/smart_find_mixin.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 import claripy
 
-from . import MemoryMixin
 from angr.errors import SimSegfaultException
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class SmartFindMixin(MemoryMixin):
