+++ angr/storage/memory_mixins/symbolic_merger_mixin.py
@@ -1,10 +1,11 @@
 from __future__ import annotations
+
 from collections.abc import Iterable
 from typing import Any
 
 import claripy
 
-from . import MemoryMixin
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class SymbolicMergerMixin(MemoryMixin):
