+++ angr/storage/memory_mixins/top_merger_mixin.py
@@ -1,8 +1,9 @@
 from __future__ import annotations
-from typing import Any
-from collections.abc import Iterable, Callable
 
-from . import MemoryMixin
+from collections.abc import Iterable, Callable
+from typing import Any
+
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class TopMergerMixin(MemoryMixin):
