+++ angr/storage/memory_mixins/unwrapper_mixin.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
-from . import MemoryMixin
+
+from angr.storage.memory_mixins.memory_mixin import MemoryMixin
 
 
 class UnwrapperMixin(MemoryMixin):
